Rename `reader` to `device_reader`
