+++ src/eeprom/device_reader.rs
