+++ src/internals.rs
@@ -2,6 +2,6 @@
 //!
 //! Anything exported by this module should be considered unstable and may change at any time.
 
-pub use crate::eeprom::reader::SiiDataProvider;
+pub use crate::eeprom::device_reader::SiiDataProvider;
 pub use crate::eeprom::EepromDataProvider;
 pub use crate::slave::slave_client::SlaveClient;
